backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
python-dotenv==1.0.1
